User config working on windows
This config directory is working on NeoVim v0.9.5 running in powershell
on windows.

This is a bare bones config at the moment.
Just enabling line wrapping and giving myself a shortcut to exit input
mode in terminal.
Also installing tetris, and a transparency shortcut.

And I changed the home screen for my ego.